serialdump: avoid signed/unsigned comparison in SLIP overflow check
index is an int while sizeof(rxbuf) is size_t, so "index >= sizeof(rxbuf)"
is a signed/unsigned comparison (caught by -Wextra). Cast the size to int;
the buffer is small and index is never negative, so the value is unchanged.

+++ tools/serial-io/serialdump.c
@@ -374,7 +374,7 @@ main(int argc, char **argv)
             }
 
             rxbuf[index++] = buf[i];
-            if(index >= sizeof(rxbuf)) {
+            if(index >= (int)sizeof(rxbuf)) {
               fprintf(stderr, "**** slip overflow\n");
               index = 0;
               flags = 2;
